project/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
numpy>=1.20,<2.0
pandas>=1.2,<2.0
scikit-learn>=0.24,<0.25
